move methods to lib/helpers.py, add tests
