+++ .github/workflows/ci.yml
@@ -15,10 +15,13 @@ jobs:
           python-version: 3.6
       
       - name: Install Ubuntu dependencies
-        run: sudo apt-get install -y -qq --no-install-recommends python3-gdal python3-pip unzip
+        run: sudo apt-get install -y -qq --no-install-recommends python3-gdal python3-pip python3-pytest unzip
 
       - name: Install Python dependencies
         run: pip install -r requirements.txt
+
+      - name: Run tests
+        run: pytest-3
 
       - name: Create output
         run: |
