+++ tiger_address_convert.py
@@ -17,6 +17,8 @@ import re
 import csv
 import math
 
+from lib.helpers import round_point, glom_all, length
+
 try:
     from osgeo import ogr
     from osgeo import osr
@@ -241,22 +243,6 @@ to_proj.SetWellKnownGeogCS( "EPSG:4326" )
 
 tr = osr.CoordinateTransformation( from_proj, to_proj )
 
-def length(segment, nodelist):
-    '''Returns the length (in feet) of a segment'''
-    first = True
-    distance = 0
-    lat_feet = 364613  #The approximate number of feet in one degree of latitude
-    for point in segment:
-        _pointid, (lat, lon) = nodelist[ round_point( point ) ]
-        if first:
-            first = False
-        else:
-            #The approximate number of feet in one degree of longitute
-            lrad = math.radians(lat)
-            lon_feet = 365527.822 * math.cos(lrad) - 306.75853 * math.cos(3 * lrad) + 0.3937 * math.cos(5 * lrad)
-            distance += math.sqrt(((lat - previous[0])*lat_feet)**2 + ((lon - previous[1])*lon_feet)**2)
-        previous = (lat, lon)
-    return distance
 
 def addressways(waylist, nodelist, first_id):
     id = first_id
@@ -497,9 +483,6 @@ def unproject( point ):
     pt = tr.TransformPoint( point[0], point[1] )
     return (pt[0], pt[1])
 
-def round_point( point, accuracy=8 ):
-    return tuple( [ round(x,accuracy) for x in point ] )
-
 def compile_nodelist( parsed_gisdata, first_id=1 ):
     nodelist = {}
 
@@ -516,74 +499,6 @@ def compile_nodelist( parsed_gisdata, first_id=1 ):
 
     return (i, nodelist)
 
-def adjacent( left, right ):
-    left_left = round_point(left[0])
-    left_right = round_point(left[-1])
-    right_left = round_point(right[0])
-    right_right = round_point(right[-1])
-
-    return ( left_left == right_left or
-             left_left == right_right or
-             left_right == right_left or
-             left_right == right_right )
-
-def glom( left, right ):
-    left = list( left )
-    right = list( right )
-
-    left_left = round_point(left[0])
-    left_right = round_point(left[-1])
-    right_left = round_point(right[0])
-    right_right = round_point(right[-1])
-
-    if left_left == right_left:
-        left.reverse()
-        return left[0:-1] + right
-
-    if left_left == right_right:
-        return right[0:-1] + left
-
-    if left_right == right_left:
-        return left[0:-1] + right
-
-    if left_right == right_right:
-        right.reverse()
-        return left[0:-1] + right
-
-    raise 'segments are not adjacent'
-
-def glom_once( segments ):
-    if len(segments)==0:
-        return segments
-
-    unsorted = list( segments )
-    x = unsorted.pop(0)
-
-    while len( unsorted ) > 0:
-        n = len( unsorted )
-
-        for i in range(0, n):
-            y = unsorted[i]
-            if adjacent( x, y ):
-                y = unsorted.pop(i)
-                x = glom( x, y )
-                break
-
-        # Sorted and unsorted lists have no adjacent segments
-        if len( unsorted ) == n:
-            break
-
-    return x, unsorted
-
-def glom_all( segments ):
-    unsorted = segments
-    chunks = []
-
-    while unsorted != []:
-        chunk, unsorted = glom_once( unsorted )
-        chunks.append( chunk )
-
-    return chunks
